silly little ongoing error refactors
