+++ app/auth/components/user-auth-form.tsx
@@ -1,6 +1,5 @@
 "use client"
 
-import { useState } from "react";
 import * as z from 'zod';
 
 import { useAuth } from "@/components/auth-provider";
@@ -10,7 +9,6 @@ import { Icons } from "@/components/ui/icons";
 import { Input } from "@/components/ui/input";
 import { cn } from "@/lib/utils";
 import { zodResolver } from "@hookform/resolvers/zod";
-import Link from "next/link";
 import { useForm } from "react-hook-form";
 // import microsoft icon
 import { BsGithub, BsMicrosoft } from "react-icons/bs";
@@ -20,8 +18,6 @@ interface UserAuthFormProps extends React.HTMLAttributes<HTMLDivElement> {
 }
 
 export function UserAuthForm({ className, signup, ...props }: UserAuthFormProps) {
-    const [authSuccessful, setAuthSuccessful] = useState<boolean>(true);
-    const [authError, setAuthError] = useState<any>(null);
     const formSchema = z.object({
         name: z.string(),
         email: z.string().email(),
@@ -37,27 +33,6 @@ export function UserAuthForm({ className, signup, ...props }: UserAuthFormProps)
         },
     })
 
-    const LoginFailure = () => {
-        return (
-            <div className='text-center'>
-                <span>Incorrect email or password. <br />Try again or </span>
-                <Link className='underline' href="/auth/signup" >
-                    sign up instead
-                </Link>
-            </div>
-        )
-    }
-
-    const SignupFailure = () => {
-        return (
-            <div className='text-center'>
-                <span>An account with this email already exists. <br /> Please </span>
-                <Link className='underline' href="/auth/login" >
-                    login instead
-                </Link>
-            </div>
-        )
-    }
 
     const OAuthButton = ({ authType, authTitle, Logo }: { authType: () => Promise<void>, authTitle: string, Logo: React.ElementType }) => {
         return (
@@ -119,26 +94,9 @@ export function UserAuthForm({ className, signup, ...props }: UserAuthFormProps)
         }
 
         if (signup) {
-            auth.createAccountEmail({ name: values.name, email: values.email, password: values.password }).then(() => {
-            }).catch((error) => {
-                if (error.code === 'auth/email-already-in-use') {
-                    setAuthError(SignupFailure);
-                } else {
-                    setAuthError("There was an error when creating the account, please try again.");
-                }
-                setAuthSuccessful(false);
-            });
+            auth.createAccountEmail({ name: values.name, email: values.email, password: values.password });
         } else {
-            auth.loginEmail({ email: values.email, password: values.password }).then(() => {
-            }).catch((error) => {
-                if (error.code === 'auth/user-not-found' || error.code === 'auth/wrong-password') {
-                    setAuthError(LoginFailure);
-                } else {
-                    setAuthError("There was an error when logging in, please try again.");
-                }
-                setAuthSuccessful(false);
-                console.log(error);
-            });
+            auth.loginEmail({ email: values.email, password: values.password });
         }
     }
 
@@ -166,9 +124,9 @@ export function UserAuthForm({ className, signup, ...props }: UserAuthFormProps)
                 </form>
             </Form>
 
-            {!authSuccessful && (
+            {auth!.error && (
                 <div className={cn("text-sm font-medium text-destructive")}>
-                    {authError}
+                    {auth!.error}
                 </div>
             )}
             <div className="relative">
@@ -185,6 +143,8 @@ export function UserAuthForm({ className, signup, ...props }: UserAuthFormProps)
             <OAuthButton authType={auth!.loginGoogle} authTitle="Google" Logo={Icons.google} />
             <OAuthButton authType={auth!.loginGithub} authTitle="Github" Logo={BsGithub} />
             <OAuthButton authType={auth!.loginMicrosoft} authTitle="Microsoft" Logo={BsMicrosoft} />
+
+            {auth!.error}
         </div>
     )
 }
